Apple Store Lookup

You can create a lookup request to search for content in the stores based on iTunes IDs, UPCs/EANs, and All Music Guide (AMG) IDs. ID-based lookups are faster and contain fewer false-positive results.

lookup_apple(
  id,
  country = NULL,
  entity = NULL,
  limit = NULL,
  sort = NULL,
  id_type = "id"
)

Arguments

id
The ID of the iTunes entity

country
The two-letter country code for the store you want to search. For a list of country codes see https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2

entity
Optional The type of results you want returned, relative to the specified media type.

limit
Optional The number of search results you want the iTunes Store to return between 1 and 200. The default is 50.

sort
Optional The order the results are returned, for most recent first select recent.

id_type
The ID type to lookup, options are:

id

The default iTunes ID

amgArtistId

AMG Artist ID

amgAlbumId

AMG Album ID

upc

UPC Album or Video ID

isbn

ISB Book ID

Value

A data.frame of any results that match the iTunes database.

If there were no successful results then it will return NULL.

See also

Examples

# Look up Jack Johnson by iTunes artist ID
lookup_apple(909253)

#>   wrapperType artistType   artistName
#> 1      artist     Artist Jack Johnson
#>                                                artistLinkUrl artistId
#> 1 https://music.apple.com/us/artist/jack-johnson/909253?uo=4   909253
#>   amgArtistId primaryGenreName primaryGenreId
#> 1      468749             Rock             21

# Look up Jack Johnson by AMG artist ID
lookup_apple(468749, id_type = "amgArtistId")

#>   wrapperType artistType   artistName
#> 1      artist     Artist Jack Johnson
#>                                                artistLinkUrl artistId
#> 1 https://music.apple.com/us/artist/jack-johnson/909253?uo=4   909253
#>   amgArtistId primaryGenreName primaryGenreId
#> 1      468749             Rock             21